One method name is not clear
The original method name is not clear enough to describe the work of the method. In fact, this method is used to search all matched results. But the original name "getDicts" may make developers misunderstand this method that is used to create dictionaries or some other works. I just would like to make it clear about this problem.